Replace unsigned long uses with size_t

+++ src/Solver/Solver.cpp
@@ -86,12 +86,12 @@ std::vector<Statement> simplifyProof(std::vector<Statement> &proof, bool isProof
     for (Statement &statement: proof) {
         statement.unused = true;
     }
-    std::vector<unsigned long> frontier;
+    std::vector<size_t> frontier;
     if (isProof) {
         frontier.push_back(proof.size() - 1);
     } else {
         std::unordered_set<char> foundVariables;
-        for (unsigned long i = proof.size() - 1; i < proof.size(); i--) {
+        for (size_t i = proof.size() - 1; i < proof.size(); i--) {
             const auto& statement = proof[i];
             if (auto variable = std::dynamic_pointer_cast<Variable>(statement.proposition)) {
                 if (!foundVariables.contains(variable->getName())) {
@@ -107,9 +107,9 @@ std::vector<Statement> simplifyProof(std::vector<Statement> &proof, bool isProof
             }
         }
     }
-    std::unordered_set<unsigned long> travelled;
+    std::unordered_set<size_t> travelled;
     while (!frontier.empty()) {
-        unsigned long index = frontier.back();
+        size_t index = frontier.back();
         frontier.pop_back();
         travelled.insert(index);
         auto& statement = proof[index];
@@ -122,14 +122,14 @@ std::vector<Statement> simplifyProof(std::vector<Statement> &proof, bool isProof
     }
 
     std::vector<Statement> out;
-    std::unordered_map<unsigned long, unsigned long> adjustedIndex;
+    std::unordered_map<size_t, size_t> adjustedIndex;
     for (int i = 0; i < proof.size(); i++) {
         const auto &statement = proof[i];
         if (statement.unused && statement.type != StatementType::PREMISE){
             continue;
         }
         adjustedIndex[i] = out.size();
-        std::vector<unsigned long> adjustedReferences;
+        std::vector<size_t> adjustedReferences;
         adjustedReferences.reserve(statement.references.size());
         for (auto reference: statement.references) {
             adjustedReferences.push_back(adjustedIndex[reference]);
@@ -171,12 +171,12 @@ std::shared_ptr<WellFormedFormula> tryBreakComplexWff(const std::shared_ptr<Well
 std::vector<Statement> solve(std::vector<Statement> argument) {
     bool isChanged = true;
     bool canBreak = true;
-    unsigned long startAt = 0;
+    size_t startAt = 0;
     bool isProven = false;
 
     verifyInitialArgument(argument);
 
-    std::vector<std::pair<unsigned long, Statement>> currentAssumptions;
+    std::vector<std::pair<size_t, Statement>> currentAssumptions;
 
     // Set up proof by contradiction
     argument.back().blocked = true;
@@ -185,7 +185,7 @@ std::vector<Statement> solve(std::vector<Statement> argument) {
             Negation::of(argument.back().proposition),
             0,
             Rule::NONE,
-            std::vector<unsigned long>{argument.size() - 1}
+            std::vector<size_t>{argument.size() - 1}
     );
     currentAssumptions.emplace_back(argument.size() - 1, argument.back());
 
@@ -196,7 +196,7 @@ std::vector<Statement> solve(std::vector<Statement> argument) {
         }
         isChanged = false;
         auto currentLength = argument.size();
-        for (unsigned long i = 0; i < currentLength; i++) {
+        for (size_t i = 0; i < currentLength; i++) {
             // Reserve space on argument to prevent reallocation and reference invalidation
             // on adding s-rule and i-rule results
             argument.reserve( argument.size() + 3 + currentLength - startAt);
@@ -212,7 +212,7 @@ std::vector<Statement> solve(std::vector<Statement> argument) {
                                           brokenProposition,
                                           leftStatement.brokenLevel,
                                           Rule::BREAK,
-                                          std::vector<unsigned long>{i});
+                                          std::vector<size_t>{i});
                     currentAssumptions.emplace_back(argument.size() - 1, argument.back());
                     canBreak = true;
                     isChanged = true;
@@ -228,12 +228,12 @@ std::vector<Statement> solve(std::vector<Statement> argument) {
                             result.proposition,
                             leftStatement.assumptionLevel,
                             result.rule,
-                            std::vector<unsigned long>{i}
+                            std::vector<size_t>{i}
                     );
                 }
             }
 
-            for (unsigned long j = std::max(i + 1, startAt); j < currentLength; j++) {
+            for (size_t j = std::max(i + 1, startAt); j < currentLength; j++) {
                 Statement& rightStatement = argument[j];
                 if (rightStatement.blocked || rightStatement.skip) {
                     continue;
@@ -247,7 +247,7 @@ std::vector<Statement> solve(std::vector<Statement> argument) {
                             Negation::of(currentAssumptions.back().second.proposition),
                             currentAssumptions.size() <= 1 ? 0 : currentAssumptions.size() - 2,
                             Rule::NONE,
-                            std::vector<unsigned long>{currentAssumptions.back().first, i, j}
+                            std::vector<size_t>{currentAssumptions.back().first, i, j}
                     );
                     currentAssumptions.pop_back();
 
@@ -282,7 +282,7 @@ std::vector<Statement> solve(std::vector<Statement> argument) {
                                 proposition.proposition,
                                 rightStatement.assumptionLevel,
                                 proposition.rule,
-                                std::vector<unsigned long>{i, j}
+                                std::vector<size_t>{i, j}
                         );
                     }
                 }
@@ -299,7 +299,7 @@ std::vector<Statement> solve(std::vector<Statement> argument) {
                                 proposition.proposition,
                                 rightStatement.assumptionLevel,
                                 proposition.rule,
-                                std::vector<unsigned long>{i, j}
+                                std::vector<size_t>{i, j}
                         );
                     }
                 }
